refactor: move slots to its own package
