+++ app/src/main/java/io/joelt/texttemplate/models/slots/Slot.kt
@@ -1,4 +1,4 @@
-package io.joelt.texttemplate.models
+package io.joelt.texttemplate.models.slots
 
 interface Slot {
     val placeholderText: String
